+++ src/converters/piral-blazor/src/events.ts
@@ -62,9 +62,11 @@ function dispatchToRoot(event: any) {
   isInternalNavigation(event) && performInternalNavigation(event);
 
   // the mutation event cannot be cloned (at least in Webkit-based browsers)
-  if (!(event instanceof MutationEvent)) {
+  if (!(event instanceof MutationEvent) && !event.processed) {
     const eventClone = new event.constructor(event.type, event);
     document.getElementById(blazorRootId)?.dispatchEvent(eventClone);
+    // make sure to only process every event once; even though multiple boundaries might be active
+    event.processed = true;
   }
 }